abstract select, status, affiliation columns
