+++ src/data/dynamic/admin/Participants.js
@@ -1,8 +1,6 @@
-import Tag from "@/components/dynamic/admin/Tag";
-import { COLORS } from "../Tags";
-import Checkbox from "@/components/dynamic/Checkbox";
 import View from "@/components/dynamic/admin/dashboards/dashboard/View";
 import { ICONS } from "./Icons";
+import { generateSelect, generateStatus } from "./Columns";
 
 export const STATUSES = {
   1: "accepted",
@@ -22,22 +20,7 @@ export const TAGS = [
 ];
 
 export const COLUMNS = [
-  {
-    id: "select",
-    width: "w-1/12",
-    header: ({ table }) => (
-      <Checkbox
-        toggle={table.getIsAllRowsSelected()}
-        onClick={table.getToggleAllRowsSelectedHandler()}
-      />
-    ),
-    cell: ({ row }) => (
-      <Checkbox
-        toggle={row.getIsSelected()}
-        onClick={row.getToggleSelectedHandler()}
-      />
-    ),
-  },
+  generateSelect(),
   {
     accessorKey: "name",
     header: "Name",
@@ -64,19 +47,7 @@ export const COLUMNS = [
     width: "w-3/12",
     cell: ({ getValue }) => <div>{getValue()}</div>,
   },
-  {
-    accessorKey: "status",
-    header: "Status",
-    width: "w-1/12",
-    enableColumnFilter: true,
-    filterFn: (row, col, filter) => {
-      const status = row.getValue(col);
-      return filter.includes(status);
-    },
-    cell: ({ getValue }) => (
-      <Tag text={STATUSES[getValue()]} color={COLORS[getValue()]} />
-    ),
-  },
+  generateStatus(STATUSES),
   {
     accessorKey: "Resume",
     header: "Resume",
