+++ src/data/dynamic/admin/Volunteers.js
@@ -1,7 +1,6 @@
 import Checkbox from "@/components/dynamic/Checkbox";
-import Tag from "@/components/dynamic/admin/Tag";
-import { COLORS } from "../Tags";
 import { AVAILABILITY } from "../form/Information";
+import { generateSelect, generateStatus } from "./Columns";
 
 export const STATUSES = {
   1: "accepted",
@@ -21,22 +20,7 @@ export const TAGS = [
 ];
 
 export const COLUMNS = [
-  {
-    id: "select",
-    width: "w-1/12",
-    header: ({ table }) => (
-      <Checkbox
-        toggle={table.getIsAllRowsSelected()}
-        onClick={table.getToggleAllRowsSelectedHandler()}
-      />
-    ),
-    cell: ({ row }) => (
-      <Checkbox
-        toggle={row.getIsSelected()}
-        onClick={row.getToggleSelectedHandler()}
-      />
-    ),
-  },
+  generateSelect(),
   {
     accessorKey: "name",
     header: "Name",
@@ -57,19 +41,7 @@ export const COLUMNS = [
     width: "w-3/12",
     cell: ({ getValue }) => <div>{getValue()}</div>,
   },
-  {
-    accessorKey: "status",
-    header: "Status",
-    width: "w-1/12",
-    enableColumnFilter: true,
-    filterFn: (row, col, filter) => {
-      const status = row.getValue(col);
-      return filter.includes(status);
-    },
-    cell: ({ getValue }) => (
-      <Tag text={STATUSES[getValue()]} color={COLORS[getValue()]} />
-    ),
-  },
+  generateStatus(STATUSES),
 ];
 
 export const DROPDOWN = ({ object }) => {
